Move trivia app to top folder
On further inspection the template Vue app was nested in a folder, with
only an additional README and config for dependabot at the top level.
As this is a temporary project we won't need dependabot to keep
dependencies up to date and the information in the top level README can
be moved. This allows us to move the app itself to the top level, which
make the project structure simpler and easier to use.