Add basic design sketches and variables
Before we start building the app, I've added some sketches to help us
figure out what the views will look like, and what components we will
need. I've also made a note of the fonts and colours we will be using,
which will come in handy once we start with styling.