Fetch questions on start button click
When the user clicks the start button, fetch questions from the API
using the session token we fetched earlier and the selected difficulty.

We will need those questions later in the questions view, so we can
store them in a store so we have easy access to them. Once that's done
we can redirect to user to the question view.

I've added the @pinia/testing as a dev dependency here, which allows
us to mock stores for tests so we can check the store actions are
called as expected.