Show question and answers on question view
On the question view, get the current question from the question store,
then render the question and the answers (in random order).

To display the HTML encoded text returned by the API correctly we could
use the v-html directive, but that would make it a lot more difficult
to manage the content of the question and answer elements. For example,
we might want to add an icon to the buttons, which would then have to
be included in the html encoded text as well. Instead we can use the
DOMParser browser API to decode the strings.

To ensure position of the correct answer is random, I've added a helper
to shuffle the array of answers. We probably don't want to shuffle
true/false answers as there the order does not reveal which answer is
correct.