Show next question and answers on button click
When the user clicks on an answer, we can increment the current
question index and display the next question and answer. We can also
keep track of the score. In a more complex app I think the score
would deserve it's own store to keep concerns separated, but for
simplicity's sake I'm storing it with the questions here.

I've also made a small update to the array shuffling helper test as of
course it is possible for the array to be in the same order after
shuffling, which would give a false negative test result. Instead we
can shuffle until we see the order change.