Reset current question index and score on restart
It was possible to go back to the home screen and start with a new set
of questions but the current question index and score would still be
saved in the store. We can avoid this by resetting them after fetching
new questions.

I've also used a variable for the base API path which I created earlier
but I had forgotten to update the fetch functions.