Add play again button to result page
On the result page, show a button that on click returns the user to the
home page so they can play again.

Returning to the home page would fetch a new session token, which meant
there was a small chance they would say a question they already
answered when they start again. Instead we can store the session token
in the store as well (which is getting a little crowded now) and only
set it if there is none. This way the user is guaranteed fresh
questions.

This can lead to running out of questions after multiple games so in
the future we could refresh the session token when receiving a session
related error code from the API.