Style questions view
We can start with the questions view, as that is where the user will
spend the most time.

The text is centred, with the question in the middle of the screen with
plenty of space around it so it is easy to focus on. The question count
is less important, so it's out of the way at the top, smaller and a
less contrasting colour. The buttons are shown in a single column on
mobile, and in two columns on desktop to make better use of the
available space.