Style result page and add button component
Style the result page so the score is shown prominently in the center
and is coloured according to how well you've done.

As we're using very similar button styling again here I thought it
would be interesting to extract it to a component. We could have also
achieved the same result with adding base button styling but using a
component allows us to use other button styles in the future.

I've used CSS modules for the styling as requested in the assignment,
but if it had been up to me I would have used Vue's scoped CSS
functionality, which seems to achieve the same result with less
complexity and is less verbose.