Style home view
Add some styling to the home view, including a nice gradient and
unusual positioning for the title to make it pop a little.

I've done some cleanup of the styles here, moving the main element
styling to the main css sheet as it was the same in each view, and
using classes for the view specific styling so they don't clash with
each other.